refactor(listen): split stats page panels
