+++ app/listen/src/pages/Stats.tsx
@@ -1,338 +1,33 @@
-import { Children, type ReactNode, useMemo, useState } from "react";
-import { ResponsiveLine } from "@nivo/line";
+import { useMemo, useState } from "react";
 import { BarChart3, Clock3, Disc3, Music2, Play, SkipForward, Tag, TrendingUp } from "lucide-react";
 import { Link } from "react-router";
 
+import {
+  OverviewCard,
+  StatsSection,
+  TopList,
+  TrendChart,
+  WindowPicker,
+} from "@/components/stats/StatsPanels";
+import {
+  buildRecapHighlights,
+  formatStatsMinutes,
+  formatStatsPercent,
+  toPlayerTrack,
+  type ReplayMix,
+  type StatsAlbum,
+  type StatsArtist,
+  type StatsGenre,
+  type StatsListResponse,
+  type StatsOverview,
+  type StatsTrack,
+  type StatsTrends,
+  type StatsWindow,
+  STATS_WINDOW_OPTIONS,
+} from "@/components/stats/stats-model";
 import { useApi } from "@/hooks/use-api";
-import { usePlayerActions, type Track } from "@/contexts/PlayerContext";
+import { usePlayerActions } from "@/contexts/PlayerContext";
 import { encPath } from "../../../shared/web/utils";
-
-type StatsWindow = "7d" | "30d" | "90d" | "365d" | "all_time";
-
-interface StatsOverview {
-  window: StatsWindow;
-  play_count: number;
-  complete_play_count: number;
-  skip_count: number;
-  minutes_listened: number;
-  active_days: number;
-  skip_rate: number;
-  top_artist: {
-    artist_name: string;
-    play_count: number;
-    minutes_listened: number;
-  } | null;
-}
-
-interface StatsTrendPoint {
-  day: string;
-  play_count: number;
-  complete_play_count: number;
-  skip_count: number;
-  minutes_listened: number;
-}
-
-interface StatsTrends {
-  window: StatsWindow;
-  points: StatsTrendPoint[];
-}
-
-interface StatsTrack {
-  track_id: number | null;
-  track_path: string | null;
-  navidrome_id?: string | null;
-  title: string;
-  artist: string;
-  album: string;
-  play_count: number;
-  complete_play_count: number;
-  minutes_listened: number;
-}
-
-interface StatsArtist {
-  artist_name: string;
-  play_count: number;
-  complete_play_count: number;
-  minutes_listened: number;
-}
-
-interface StatsAlbum {
-  artist: string;
-  album: string;
-  play_count: number;
-  complete_play_count: number;
-  minutes_listened: number;
-}
-
-interface StatsGenre {
-  genre_name: string;
-  play_count: number;
-  complete_play_count: number;
-  minutes_listened: number;
-}
-
-interface StatsListResponse<T> {
-  window: StatsWindow;
-  items: T[];
-}
-
-interface ReplayMix {
-  window: StatsWindow;
-  title: string;
-  subtitle: string;
-  track_count: number;
-  minutes_listened: number;
-  items: StatsTrack[];
-}
-
-interface RecapHighlight {
-  title: string;
-  body: string;
-}
-
-const WINDOW_OPTIONS: { value: StatsWindow; label: string }[] = [
-  { value: "7d", label: "7D" },
-  { value: "30d", label: "30D" },
-  { value: "90d", label: "90D" },
-  { value: "365d", label: "1Y" },
-  { value: "all_time", label: "All time" },
-];
-
-function formatMinutes(minutes: number): string {
-  if (!Number.isFinite(minutes) || minutes <= 0) return "0m";
-  if (minutes >= 60) {
-    const hours = Math.floor(minutes / 60);
-    const remaining = Math.round(minutes % 60);
-    return remaining > 0 ? `${hours}h ${remaining}m` : `${hours}h`;
-  }
-  return `${Math.round(minutes)}m`;
-}
-
-function formatPercent(value: number): string {
-  return `${Math.round((value || 0) * 100)}%`;
-}
-
-function OverviewCard({
-  icon: Icon,
-  label,
-  value,
-  hint,
-}: {
-  icon: React.ComponentType<{ size?: number; className?: string }>;
-  label: string;
-  value: string;
-  hint?: string;
-}) {
-  return (
-    <div className="rounded-3xl border border-white/10 bg-white/[0.03] p-5">
-      <div className="flex items-start justify-between gap-3">
-        <div>
-          <p className="text-[11px] font-bold uppercase tracking-[0.2em] text-white/35">{label}</p>
-          <p className="mt-3 text-2xl font-bold text-foreground">{value}</p>
-          {hint ? <p className="mt-2 text-sm text-muted-foreground">{hint}</p> : null}
-        </div>
-        <div className="flex h-10 w-10 items-center justify-center rounded-2xl border border-primary/15 bg-primary/10 text-primary">
-          <Icon size={18} />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function Section({
-  title,
-  subtitle,
-  children,
-}: {
-  title: string;
-  subtitle?: string;
-  children: ReactNode;
-}) {
-  return (
-    <section className="rounded-3xl border border-white/10 bg-white/[0.03] p-5 sm:p-6">
-      <div className="mb-5">
-        <h2 className="text-lg font-semibold text-foreground">{title}</h2>
-        {subtitle ? <p className="mt-1 text-sm text-muted-foreground">{subtitle}</p> : null}
-      </div>
-      {children}
-    </section>
-  );
-}
-
-function WindowPicker({
-  value,
-  onChange,
-}: {
-  value: StatsWindow;
-  onChange: (value: StatsWindow) => void;
-}) {
-  return (
-    <div className="inline-flex rounded-2xl border border-white/10 bg-white/[0.03] p-1">
-      {WINDOW_OPTIONS.map((option) => (
-        <button
-          key={option.value}
-          onClick={() => onChange(option.value)}
-          className={`rounded-xl px-3 py-1.5 text-sm transition-colors ${
-            value === option.value
-              ? "bg-primary text-primary-foreground"
-              : "text-white/50 hover:text-white hover:bg-white/5"
-          }`}
-        >
-          {option.label}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-function TopList({
-  title,
-  emptyText,
-  loading = false,
-  children,
-}: {
-  title: string;
-  emptyText: string;
-  loading?: boolean;
-  children: ReactNode;
-}) {
-  const hasVisibleItems = Children.count(children) > 0;
-
-  return (
-    <div className="rounded-2xl border border-white/10 bg-black/10 p-4">
-      <h3 className="text-sm font-semibold text-foreground">{title}</h3>
-      <div className="mt-3 space-y-2">
-        {loading ? <p className="text-sm text-muted-foreground">Loading...</p> : hasVisibleItems ? children : <p className="text-sm text-muted-foreground">{emptyText}</p>}
-      </div>
-    </div>
-  );
-}
-
-function TrendChart({ points, loading }: { points: StatsTrendPoint[]; loading?: boolean }) {
-  const data = useMemo(() => [
-    {
-      id: "Minutes",
-      data: points.map((point) => ({
-        x: point.day,
-        y: Number(point.minutes_listened.toFixed(2)),
-      })),
-    },
-  ], [points]);
-
-  if (loading) {
-    return (
-      <div className="flex h-72 items-center justify-center rounded-2xl border border-dashed border-white/10 bg-black/10 text-sm text-muted-foreground">
-        Loading trend data...
-      </div>
-    );
-  }
-
-  if (points.length === 0) {
-    return (
-      <div className="flex h-72 items-center justify-center rounded-2xl border border-dashed border-white/10 bg-black/10 text-sm text-muted-foreground">
-        Start listening and your daily curve will appear here.
-      </div>
-    );
-  }
-
-  return (
-    <div className="h-72 rounded-2xl border border-white/10 bg-black/10 p-3">
-      <ResponsiveLine
-        data={data}
-        margin={{ top: 20, right: 20, bottom: 40, left: 50 }}
-        xScale={{ type: "point" }}
-        yScale={{ type: "linear", min: 0, max: "auto", stacked: false, reverse: false }}
-        axisTop={null}
-        axisRight={null}
-        colors={["#22d3ee"]}
-        enableGridX={false}
-        pointSize={7}
-        pointColor="#22d3ee"
-        pointBorderWidth={0}
-        useMesh
-        theme={{
-          text: { fill: "rgba(255,255,255,0.45)", fontSize: 11 },
-          axis: {
-            ticks: { text: { fill: "rgba(255,255,255,0.35)" } },
-            legend: { text: { fill: "rgba(255,255,255,0.35)" } },
-            domain: { line: { stroke: "rgba(255,255,255,0.08)" } },
-          },
-          grid: { line: { stroke: "rgba(255,255,255,0.06)" } },
-          crosshair: { line: { stroke: "rgba(255,255,255,0.2)", strokeWidth: 1 } },
-          tooltip: {
-            container: {
-              background: "#0f1117",
-              color: "#fff",
-              border: "1px solid rgba(255,255,255,0.08)",
-              borderRadius: "14px",
-            },
-          },
-        }}
-        axisBottom={{
-          tickRotation: points.length > 14 ? -45 : 0,
-          format: (value) => {
-            const date = new Date(`${String(value)}T12:00:00`);
-            return date.toLocaleDateString("en-US", { month: "short", day: "numeric" });
-          },
-        }}
-        axisLeft={{
-          format: (value) => `${Math.round(Number(value))}m`,
-        }}
-        tooltip={({ point }) => (
-          <div className="px-3 py-2">
-            <div className="text-xs font-semibold text-white">{String(point.data.xFormatted)}</div>
-            <div className="mt-1 text-sm text-cyan-300">{point.data.yFormatted} minutes</div>
-          </div>
-        )}
-      />
-    </div>
-  );
-}
-
-function buildRecapHighlights(
-  overview: StatsOverview | undefined,
-  replay: ReplayMix | undefined,
-  topArtists: StatsArtist[],
-  topTracks: StatsTrack[],
-): RecapHighlight[] {
-  const highlights: RecapHighlight[] = [];
-
-  if (overview?.top_artist?.artist_name) {
-    highlights.push({
-      title: `${overview.top_artist.artist_name} led this window`,
-      body: `${overview.top_artist.play_count} plays and ${formatMinutes(overview.top_artist.minutes_listened)} listened.`,
-    });
-  }
-
-  if (topTracks[0] && topTracks[0].play_count > 0) {
-    highlights.push({
-      title: `"${topTracks[0].title}" kept coming back`,
-      body: `${topTracks[0].artist} · ${topTracks[0].play_count} plays in this window.`,
-    });
-  }
-
-  if (overview && overview.play_count > 0) {
-    const cadence =
-      overview.active_days >= 20
-        ? "You've been listening almost every day."
-        : overview.active_days >= 10
-          ? "This window has had a steady rhythm."
-          : "This window is still taking shape.";
-    highlights.push({
-      title: `${formatMinutes(overview.minutes_listened)} listened`,
-      body: `${cadence} ${overview.complete_play_count} completed plays so far.`,
-    });
-  }
-
-  if (replay?.track_count && replay.track_count > 0) {
-    highlights.push({
-      title: `${replay.track_count} tracks define this replay`,
-      body: `${topArtists.length ? `Spread across ${Math.min(topArtists.length, 8)} key artists.` : "A first replay object is ready to play."}`,
-    });
-  }
-
-  return highlights.slice(0, 3);
-}
 
 export function Stats() {
   const [selectedWindow, setSelectedWindow] = useState<StatsWindow>("30d");
@@ -354,16 +49,6 @@ export function Stats() {
     () => buildRecapHighlights(overview ?? undefined, replay ?? undefined, topArtistItems, topTrackItems),
     [overview, replay, topArtistItems, topTrackItems],
   );
-
-  const toPlayerTrack = (item: StatsTrack): Track => ({
-    id: item.track_path || String(item.track_id || `${item.artist}-${item.title}`),
-    title: item.title,
-    artist: item.artist,
-    album: item.album,
-    path: item.track_path || undefined,
-    libraryTrackId: item.track_id || undefined,
-    navidromeId: item.navidrome_id || undefined,
-  });
 
   const playTopTrack = (item: StatsTrack) => {
     const track = toPlayerTrack(item);
@@ -398,24 +83,24 @@ export function Stats() {
       </div>
 
       <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
-        <OverviewCard
-          icon={Clock3}
-          label="Time listened"
-          value={overview ? formatMinutes(overview.minutes_listened) : overviewLoading ? "..." : "0m"}
-          hint={overview ? `${overview.active_days} active days` : "Listening time in the selected window"}
-        />
+          <OverviewCard
+            icon={Clock3}
+            label="Time listened"
+            value={overview ? formatStatsMinutes(overview.minutes_listened) : overviewLoading ? "..." : "0m"}
+            hint={overview ? `${overview.active_days} active days` : "Listening time in the selected window"}
+          />
         <OverviewCard
           icon={Music2}
           label="Qualified plays"
           value={overview ? String(overview.play_count) : overviewLoading ? "..." : "0"}
           hint={overview ? `${overview.complete_play_count} completed plays` : "Valid plays recorded"}
         />
-        <OverviewCard
-          icon={SkipForward}
-          label="Skip rate"
-          value={overview ? formatPercent(overview.skip_rate) : overviewLoading ? "..." : "0%"}
-          hint={overview ? `${overview.skip_count} skips` : "Tracks you moved on from"}
-        />
+          <OverviewCard
+            icon={SkipForward}
+            label="Skip rate"
+            value={overview ? formatStatsPercent(overview.skip_rate) : overviewLoading ? "..." : "0%"}
+            hint={overview ? `${overview.skip_count} skips` : "Tracks you moved on from"}
+          />
         <OverviewCard
           icon={TrendingUp}
           label="Top artist"
@@ -424,7 +109,7 @@ export function Stats() {
         />
       </div>
 
-      <Section
+      <StatsSection
         title={replay?.title || "Replay"}
         subtitle={replay?.subtitle || "Turn this listening window into a playable recap."}
       >
@@ -436,11 +121,13 @@ export function Stats() {
             </div>
             <div className="rounded-2xl border border-white/10 bg-black/10 px-4 py-3">
               <div className="text-[10px] uppercase tracking-[0.16em] text-white/35">Minutes</div>
-              <div className="mt-1 text-lg font-semibold text-foreground">{formatMinutes(replay?.minutes_listened ?? 0)}</div>
+              <div className="mt-1 text-lg font-semibold text-foreground">{formatStatsMinutes(replay?.minutes_listened ?? 0)}</div>
             </div>
             <div className="rounded-2xl border border-white/10 bg-black/10 px-4 py-3">
               <div className="text-[10px] uppercase tracking-[0.16em] text-white/35">Window</div>
-              <div className="mt-1 text-lg font-semibold text-foreground">{WINDOW_OPTIONS.find((item) => item.value === selectedWindow)?.label ?? selectedWindow}</div>
+              <div className="mt-1 text-lg font-semibold text-foreground">
+                {STATS_WINDOW_OPTIONS.find((item) => item.value === selectedWindow)?.label ?? selectedWindow}
+              </div>
             </div>
           </div>
 
@@ -481,9 +168,9 @@ export function Stats() {
             Keep listening and your replay object will start to take shape.
           </div>
         )}
-      </Section>
+      </StatsSection>
 
-      <Section
+      <StatsSection
         title="Your window so far"
         subtitle="A more readable summary of what this period says about your listening."
       >
@@ -502,17 +189,17 @@ export function Stats() {
             </div>
           )}
         </div>
-      </Section>
+      </StatsSection>
 
-      <Section
+      <StatsSection
         title="Daily trend"
         subtitle="Your listening curve across the selected time window."
       >
         <TrendChart points={trends?.points ?? []} loading={trendsLoading} />
-      </Section>
+      </StatsSection>
 
       <div className="grid gap-4 xl:grid-cols-2">
-        <Section
+        <StatsSection
           title="Top tracks"
           subtitle="The songs that defined this window."
         >
@@ -532,14 +219,14 @@ export function Stats() {
                 </div>
                 <div className="shrink-0 text-right">
                   <div className="text-sm font-medium text-foreground">{item.play_count}</div>
-                  <div className="text-[11px] text-muted-foreground">{formatMinutes(item.minutes_listened)}</div>
+                  <div className="text-[11px] text-muted-foreground">{formatStatsMinutes(item.minutes_listened)}</div>
                 </div>
               </button>
             ))}
           </TopList>
-        </Section>
+        </StatsSection>
 
-        <Section
+        <StatsSection
           title="Top artists"
           subtitle="Who you kept coming back to."
         >
@@ -555,17 +242,17 @@ export function Stats() {
                 </div>
                 <div className="min-w-0 flex-1">
                   <div className="truncate text-sm font-medium text-foreground">{item.artist_name}</div>
-                  <div className="text-xs text-muted-foreground">{formatMinutes(item.minutes_listened)}</div>
+                  <div className="text-xs text-muted-foreground">{formatStatsMinutes(item.minutes_listened)}</div>
                 </div>
                 <div className="text-sm font-medium text-foreground">{item.play_count}</div>
               </Link>
             ))}
           </TopList>
-        </Section>
+        </StatsSection>
       </div>
 
       <div className="grid gap-4 xl:grid-cols-2">
-        <Section
+        <StatsSection
           title="Top albums"
           subtitle="Records that shaped the window."
         >
@@ -585,14 +272,14 @@ export function Stats() {
                 </div>
                 <div className="text-right">
                   <div className="text-sm font-medium text-foreground">{item.play_count}</div>
-                  <div className="text-[11px] text-muted-foreground">{formatMinutes(item.minutes_listened)}</div>
+                  <div className="text-[11px] text-muted-foreground">{formatStatsMinutes(item.minutes_listened)}</div>
                 </div>
               </Link>
             ))}
           </TopList>
-        </Section>
+        </StatsSection>
 
-        <Section
+        <StatsSection
           title="Top genres"
           subtitle="Your strongest stylistic pull in this window."
         >
@@ -607,13 +294,13 @@ export function Stats() {
                 </div>
                 <div className="min-w-0 flex-1">
                   <div className="truncate text-sm font-medium text-foreground">{item.genre_name}</div>
-                  <div className="text-xs text-muted-foreground">{formatMinutes(item.minutes_listened)}</div>
+                  <div className="text-xs text-muted-foreground">{formatStatsMinutes(item.minutes_listened)}</div>
                 </div>
                 <div className="text-sm font-medium text-foreground">{item.play_count}</div>
               </div>
             ))}
           </TopList>
-        </Section>
+        </StatsSection>
       </div>
 
       {allSectionsLoaded && !overview?.play_count ? (
